ai_service/requirements.txt
ultralytics==8.4.33
opencv-python==4.13.0.92
torch==2.11.0
torchvision==0.26.0
numpy==2.4.4
fastapi==0.135.3
uvicorn==0.42.0
python-socketio[client]==5.16.1
requests==2.33.1
python-dotenv==1.2.2
pydantic==2.12.5
psutil==7.2.2